fix: abstract and test generate triples for nonprofit
