+++ apps/web/core/utils/contracts/generate-triples-for-nonprofit.test.ts
@@ -1,0 +1,9 @@
+import { describe, expect, it } from 'vitest';
+
+import { generateTriplesForNonprofit } from './generate-triples-for-nonprofit';
+
+describe('Generating triples in a nonprofit space', () => {
+  it('generates initial content template', () => {
+    expect(generateTriplesForNonprofit('spaceConfigEntityId', 'spaceName', 'spaceAddress')).toMatchSnapshot();
+  });
+});
